+++ extensions/unicode-symbols/src/components/CharacterActionPanel.tsx
@@ -13,9 +13,18 @@ export const CharacterActionPanel = ({ item, section }: { item: Character; secti
   const { findHtmlEntity, setDatasetFilterAnd, filter } = useListContext();
   const html = findHtmlEntity(item.c);
 
-  const { addToRecentlyUsedItems, isRecentlyUsed, clearRecentlyUsedItems, removeFromRecentlyUsedItems } =
-    useListContext();
+  const {
+    addToRecentlyUsedItems,
+    isRecentlyUsed,
+    clearRecentlyUsedItems,
+    removeFromRecentlyUsedItems,
+    addToFavorites,
+    removeFromFavorites,
+    clearFavorites,
+    isFavorite,
+  } = useListContext();
   const recentlyUsed = useMemo(() => isRecentlyUsed(item), [isRecentlyUsed, item]);
+  const isItemFavorite = useMemo(() => isFavorite(item), [isFavorite, item]);
 
   const copyAction = useMemo(() => {
     return (
@@ -23,6 +32,7 @@ export const CharacterActionPanel = ({ item, section }: { item: Character; secti
         title="Copy Character to Clipboard"
         content={item.v}
         onCopy={() => addToRecentlyUsedItems(item)}
+        shortcut={{ modifiers: ["cmd"], key: "c" }}
       />
     );
   }, [item, addToRecentlyUsedItems]);
@@ -34,6 +44,7 @@ export const CharacterActionPanel = ({ item, section }: { item: Character; secti
         content={item.v}
         icon={frontmostApp ? { fileIcon: frontmostApp.path } : Icon.Clipboard}
         onPaste={() => addToRecentlyUsedItems(item)}
+        shortcut={{ modifiers: ["cmd"], key: "v" }}
       />
     );
   }, [frontmostApp, item, addToRecentlyUsedItems]);
@@ -58,25 +69,30 @@ export const CharacterActionPanel = ({ item, section }: { item: Character; secti
   return (
     <ActionPanel>
       <ActionPanel.Section title="Main">{main}</ActionPanel.Section>
-      <ActionPanel.Section title="Hex">
+      <ActionPanel.Section title="Formats">
         <Action.CopyToClipboard
           title={`Copy "${numberToHex(item.c)}" (HEX) to Clipboard`}
           content={numberToHex(item.c)}
           onCopy={() => addToRecentlyUsedItems(item)}
           shortcut={{ modifiers: ["cmd"], key: "h" }}
         />
-      </ActionPanel.Section>
-      <ActionPanel.Section title="HTML">
+        <Action.CopyToClipboard
+          // eslint-disable-next-line @raycast/prefer-title-case
+          title={`Copy "\\u${numberToHex(item.c)}" (Unicode Escape) to Clipboard`}
+          content={`\\u${numberToHex(item.c)}`}
+          onCopy={() => addToRecentlyUsedItems(item)}
+          shortcut={{ modifiers: ["cmd"], key: "u" }}
+        />
         {html ? (
           <Action.CopyToClipboard
-            title={`Copy "${html}" (HTML) to Clipboard`}
+            title={`Copy "${html}" (HTML Entity) to Clipboard`}
             content={html}
             onCopy={() => addToRecentlyUsedItems(item)}
             shortcut={{ modifiers: ["cmd", "shift"], key: "h" }}
           />
         ) : null}
         <Action.CopyToClipboard
-          title={`Copy "&#${item.c};" (HTML) to Clipboard`}
+          title={`Copy "&#${item.c};" (HTML Decimal) to Clipboard`}
           content={`&#${item.c};`}
           onCopy={() => addToRecentlyUsedItems(item)}
           shortcut={{ modifiers: ["cmd", "shift"], key: html !== null ? "t" : "h" }}
@@ -89,17 +105,42 @@ export const CharacterActionPanel = ({ item, section }: { item: Character; secti
               title={`Set Filter to "${section}"`}
               icon={Icon.Filter}
               onAction={() => setDatasetFilterAnd(section)}
+              shortcut={{ modifiers: ["cmd"], key: "f" }}
             />
           )}
           {filter !== null && (
             <Action
-              title={`Set Filter to "<All Characters>"`}
-              icon={Icon.Filter}
+              title={`Clear Filter (Show All Characters)`}
+              icon={Icon.XMarkCircle}
               onAction={() => setDatasetFilterAnd(null)}
+              shortcut={{ modifiers: ["cmd", "shift"], key: "f" }}
             />
           )}
         </ActionPanel.Section>
       )}
+      <ActionPanel.Section title="Favorites">
+        {isItemFavorite ? (
+          <Action
+            title="Remove from Favorites"
+            icon={Icon.HeartDisabled}
+            onAction={() => removeFromFavorites(item)}
+            shortcut={{ modifiers: ["cmd"], key: "l" }}
+          />
+        ) : (
+          <Action
+            title="Add to Favorites"
+            icon={Icon.Heart}
+            onAction={() => addToFavorites(item)}
+            shortcut={{ modifiers: ["cmd"], key: "l" }}
+          />
+        )}
+        <Action
+          title="Clear All Favorites"
+          icon={Icon.Trash}
+          onAction={() => clearFavorites()}
+          shortcut={{ modifiers: ["cmd", "shift"], key: "f" }}
+        />
+      </ActionPanel.Section>
       <ActionPanel.Section title="Recently Used">
         {recentlyUsed ? (
           <>
@@ -107,8 +148,14 @@ export const CharacterActionPanel = ({ item, section }: { item: Character; secti
               title="Remove from Recently Used"
               icon={Icon.Trash}
               onAction={() => removeFromRecentlyUsedItems(item)}
+              shortcut={{ modifiers: ["cmd"], key: "j" }}
             />
-            <Action title="Remove All from Recently Used" icon={Icon.Trash} onAction={() => clearRecentlyUsedItems()} />
+            <Action
+              title="Clear All Recently Used"
+              icon={Icon.Trash}
+              onAction={() => clearRecentlyUsedItems()}
+              shortcut={{ modifiers: ["cmd", "shift"], key: "j" }}
+            />
           </>
         ) : null}
       </ActionPanel.Section>
@@ -116,10 +163,12 @@ export const CharacterActionPanel = ({ item, section }: { item: Character; secti
         <Action.OpenInBrowser
           title="Open Character on Compart"
           url={`https://www.compart.com/en/unicode/U+${numberToHex(item.c)}`}
+          shortcut={{ modifiers: ["cmd"], key: "o" }}
         />
         <Action.OpenInBrowser
           title="Open Character on Unicode Explorer"
           url={`https://unicode-explorer.com/c/${numberToHex(item.c)}`}
+          shortcut={{ modifiers: ["cmd", "shift"], key: "o" }}
         />
       </ActionPanel.Section>
     </ActionPanel>